requirements.txt
numpy>=1.19.0
opencv-python>=4.5.0
torch>=1.9.0
torchvision>=0.10.0
matplotlib>=3.3.0
pillow>=8.0.0
pyzbar>=0.1.8
scikit-learn>=0.24.0
tqdm>=4.50.0
pyyaml>=5.4.0